requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
jsonschema==3.2.0
requests==2.23.0
python-Levenshtein==0.12.0